[Gold V] Title: 피타고라스 정리의 증명, Time: 348 ms, Memory: 2020 KB, Score: 100 point -BaekjoonHub

+++ 백준/Gold/33466. 피타고라스 정리의 증명/피타고라스 정리의 증명.cc
@@ -24,9 +24,9 @@ int main() {
 
 	int T; cin >> T;
 	while (T--) {
-		int N; cin >> N;
+		ll N; cin >> N;
 
-		int res = 0;
+		ll res = 0;
 		res += N; // (1,1), (2,2), ... (N,N)
 		res += N/2*2; // {(1,2), (2,4), ... (N/2, N/2*2)} * 2
